requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic[email]==2.7.1
supabase==2.9.1
gotrue==2.9.0
httpx==0.27.0
python-dotenv==1.0.1
numpy>=1.24.0
scipy>=1.11.0
opencv-python-headless>=4.8.0